edible cockroaches

+++ code/modules/reagents/reagent_containers/food/snacks/meat.dm
@@ -162,7 +162,7 @@
 
 
 /obj/item/weapon/reagent_containers/food/snacks/rawcrab
-	name = "raw crab meat"
+	name = "crab meat"
 	desc = "Fresh crab meat. Looks tasty."
 	icon_state = "raw_crabmeat"
 	health = 180
@@ -199,5 +199,36 @@
 					new/mob/living/simple_animal/cockroach(get_turf(src))
 				else
 					new/mob/living/simple_animal/fly(get_turf(src))
+		spawn(3600)
+			qdel(src)
+
+
+/obj/item/weapon/reagent_containers/food/snacks/cockroach
+	name = "cockroach"
+	desc = "A dead cockroach. No, please don't make me eat it..."
+	icon_state = "cockroach"
+	health = 180
+	filling_color = "#773B00"
+	raw = TRUE
+	var/rotten = FALSE
+	non_vegetarian = TRUE
+	satisfaction = -20
+	New()
+		..()
+		reagents.add_reagent("protein", 1)
+		reagents.add_reagent("food_poisoning", 1)
+		bitesize = 3
+
+
+/obj/item/weapon/reagent_containers/food/snacks/cockroach/New()
+	..()
+	spawn(3000) //5 minutes
+		if (!src)
+			return
+		icon_state = "rotten_cockroach"
+		name = "rotten [name]"
+		if (reagents)
+			reagents.add_reagent("food_poisoning", 1)
+		rotten = TRUE
 		spawn(3600)
 			qdel(src)
